[TASK] Fix phpstan checkFunctionArgumentTypes errors in ext:workspaces
This patch fixes incompatible type usage in function arguments
and is preparatory work for introducing native type hints and
strict mode in all core files.

Releases: master, 10.4
Resolves: #92106
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65592
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>

+++ typo3/sysext/workspaces/Classes/Preview/PreviewUriBuilder.php
@@ -143,7 +143,7 @@ class PreviewUriBuilder
     {
         $movePlaceholder = [];
         if ($table === 'pages') {
-            return BackendUtility::viewOnClick(BackendUtility::getLiveVersionIdOfRecord('pages', $uid));
+            return BackendUtility::viewOnClick((int)BackendUtility::getLiveVersionIdOfRecord('pages', $uid));
         }
 
         if ($liveRecord === null) {
